add transformer encoder
